requirements.txt
# Core runtime
python-telegram-bot>=20.3,<21
python-dotenv>=1.0
openai>=1.0

# LangChain stack (RAG)
langchain>=0.2
langchain-core>=0.2
langchain-openai>=0.1.7
langchain-community>=0.2

# Vector store backend
faiss-cpu>=1.7.4; platform_system != "Windows"

# Database (PostgreSQL)
psycopg2-binary>=2.9

# Dashboard (Flask app)
Flask>=3.0,<4
gunicorn>=21.2; platform_system != "Windows"

